requirements.txt
# Web Framework
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
pydantic-settings~=2.11.0

# LangChain & AI
langchain==0.3.27
langchain-community~=0.3.30
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langchain-google-genai==2.1.0
langgraph~=0.6.7

# OpenAI & Embeddings
openai==1.109.1

# Database
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg==3.2.10
psycopg-pool==3.2.6
pgvector==0.3.6
greenlet==3.1.1

# HTML Parsing
beautifulsoup4==4.13.5

# HTTP Client
aiohttp==3.11.16

# Environment
python-dotenv==1.1.1
bs4~=0.0.2
requests~=2.32.5
